Flock now chases the player, NaN bug fixed
Created the framework for a simple game: Attack of the Boids. Player needs to move to avoid a flock of boids that's constantly chasing them. For some reason the boids' velocity would be set to NaN under certain conditions, but that was fixed with a simple check in the Agent.cpp code.